add client and server (#3)
* not have working code

* make most work

Still need work on the terminal log

* fix logic comparison

There was a problem with that even though a bid was accepted the
client thought it was a fail.
The fault was in client line 97. Before it said resp and had a to
string method. The fix was to write resp.GetResponse.
The old way would resturn "response: "Success"".